refactor: trim hardware-profiles-shared to truly multi-consumer items
Move host-only files back to frontend/src/concepts/hardwareProfiles/:
- HardwareProfilesContext, useHardwareProfilesStatusProvider,
  useServingHardwareProfileConfig, useModelServingPodSpecOptionsState,
  kueueConstants

Split mixed files (types.ts, const.ts, utils.ts) so shared versions
contain only multi-consumer items and frontend versions re-export from
shared + define host-only items locally.

Internal deps of multi-consumer components (HardwareProfileSelect,
HardwareProfileCustomize, HardwareProfileDetailsPopover,
HardwareProfileBindingStateLabel) remain in shared since they are
required by HardwareProfileFormSection and HardwareProfileTableColumn
which are consumed by model-serving.
