+++ frontend/src/concepts/hardwareProfiles/HardwareProfileSelect.tsx
@@ -1,2 +1,2 @@
-// eslint-disable-next-line @odh-dashboard/no-restricted-imports -- re-exporting from hardware-profiles for backward compatibility
+// eslint-disable-next-line @odh-dashboard/no-restricted-imports -- re-exporting from hardware-profiles-shared for backward compatibility
 export { HardwareProfileSelect as default } from '@odh-dashboard/hardware-profiles-shared';
