--- packages/hardware-profiles-shared/src/kueueConstants.ts
@@ -1,7 +1,0 @@
-/**
- * Constants for Kueue-related messages
- */
-export const KUEUE_MODEL_DEPLOYMENT_DISABLED_MESSAGE =
-  'Model deployment requires Kueue. Contact your admin.';
-export const KUEUE_WORKBENCH_CREATION_DISABLED_MESSAGE =
-  'Workbench creation requires Kueue. Contact your admin.';
