+++ packages/hardware-profiles-shared/src/types.ts
@@ -7,32 +7,14 @@ import {
   Toleration,
   NodeSelector,
 } from '@odh-dashboard/k8s-core';
-import type { AcceleratorProfileKind } from '@odh-dashboard/internal/k8sTypes';
-import type useAcceleratorProfileFormState from '@odh-dashboard/internal/utilities/useAcceleratorProfileFormState';
 import { HardwareProfileBindingState } from './const';
 import { useHardwareProfileConfig } from './useHardwareProfileConfig';
-
-export type WarningNotification = {
-  title: string;
-  message: string;
-};
-
-export enum HardwareProfileWarningType {
-  HARDWARE_PROFILES_MISSING_CPU_MEMORY = 'hardware_profiles_missing_cpu_memory',
-  CANNOT_BE_NEGATIVE = 'cannot_be_negative',
-  CANNOT_BE_DECIMAL = 'cannot_be_decimal',
-  INVALID_UNIT = 'invalid_unit',
-  INVALID_NO = 'invalid_no',
-  MISSING_VALUE = 'missing_value',
-  OUT_OF_RANGE = 'out_of_range',
-  OTHER = 'other',
-}
 
 export type PodSpecOptions = {
   resources?: ContainerResources;
   tolerations?: Toleration[];
   nodeSelector?: NodeSelector;
-  selectedAcceleratorProfile?: AcceleratorProfileKind;
+  selectedAcceleratorProfile?: unknown;
   selectedHardwareProfile?: HardwareProfileKind;
 };
 
@@ -41,17 +23,6 @@ export type HardwarePodSpecOptions = {
   tolerations?: Toleration[];
   nodeSelector?: NodeSelector;
   selectedHardwareProfile?: HardwareProfileKind;
-};
-
-/**
- * @deprecated
- * only in modelmesh deprecation path
- * modelmesh: RHOAIENG-34917, RHOAIENG-19185
- */
-export type PodSpecOptionsAcceleratorState<T extends PodSpecOptions> = {
-  acceleratorProfile: ReturnType<typeof useAcceleratorProfileFormState>;
-  hardwareProfile: ReturnType<typeof useHardwareProfileConfig>;
-  podSpecOptions: T;
 };
 
 export type HardwarePodSpecOptionsState<T extends HardwarePodSpecOptions> = {
