--- packages/hardware-profiles-shared/src/useModelServingPodSpecOptionsState.ts
@@ -1,99 +1,0 @@
-/**
- * @jest-environment jsdom
- */
-import React from 'react';
-import type { InferenceServiceKind, ServingRuntimeKind } from '@odh-dashboard/internal/k8sTypes';
-import { useAppContext } from '@odh-dashboard/internal/app/AppContext';
-import { getModelServingSizes } from '@odh-dashboard/internal/concepts/modelServing/modelServingSizesUtils';
-import { useDeepCompareMemoize } from '@odh-dashboard/internal/utilities/useDeepCompareMemoize';
-import type { ModelServingSize } from '@odh-dashboard/internal/pages/modelServing/screens/types';
-import { getInferenceServiceSize } from '@odh-dashboard/internal/pages/modelServing/utils';
-import useServingHardwareProfileConfig from './useServingHardwareProfileConfig';
-import { PodSpecOptions, HardwarePodSpecOptionsState } from './types';
-
-/**
- * most of this file (everything that uses and descends from PodSpecOptions) is deprecated
- * modelmesh: RHOAIENG-34917, RHOAIENG-19185
- *
- * when modelmesh is removed; remove everything that uses and descends from PodSpecOptions
- * (do this last; it should be the last thing removed and be obvious afterwards)
- */
-export type ModelServingPodSpecOptions = PodSpecOptions & {
-  selectedModelSize?: ModelServingSize;
-};
-
-export type ModelServingSizeState = {
-  sizes: ModelServingSize[];
-  selectedSize: ModelServingSize;
-  setSelectedSize: (modelSize: ModelServingSize) => void;
-};
-
-export type ModelServingHardwareProfileState =
-  HardwarePodSpecOptionsState<ModelServingPodSpecOptions> & {
-    modelSize: ModelServingSizeState;
-  };
-
-export const useModelServingHardwareProfileState = (
-  servingRuntime?: ServingRuntimeKind,
-  inferenceService?: InferenceServiceKind,
-): ModelServingHardwareProfileState => {
-  const { dashboardConfig } = useAppContext();
-  const sizes = useDeepCompareMemoize(getModelServingSizes(dashboardConfig));
-  const existingSize = useDeepCompareMemoize(
-    getInferenceServiceSize(sizes, inferenceService, servingRuntime),
-  );
-  const [modelSize, setModelSize] = React.useState(sizes[0]);
-
-  // set initial model size
-  React.useEffect(() => {
-    if (inferenceService) {
-      setModelSize(existingSize);
-    }
-  }, [inferenceService, existingSize]);
-
-  const hardwareProfile = useServingHardwareProfileConfig(inferenceService);
-
-  let podSpecOptions: ModelServingPodSpecOptions = {
-    resources: {},
-    tolerations: undefined,
-    nodeSelector: undefined,
-  };
-
-  const existingResources =
-    inferenceService?.spec.predictor.model?.resources ||
-    servingRuntime?.spec.containers[0].resources ||
-    {};
-  const existingTolerations =
-    inferenceService?.spec.predictor.tolerations || servingRuntime?.spec.tolerations;
-  const existingNodeSelector =
-    inferenceService?.spec.predictor.nodeSelector || servingRuntime?.spec.nodeSelector;
-
-  const annotationData = {
-    selectedHardwareProfile: hardwareProfile.formData.selectedProfile,
-  };
-  if (hardwareProfile.formData.useExistingSettings) {
-    podSpecOptions = {
-      resources: existingResources,
-      tolerations: existingTolerations,
-      nodeSelector: existingNodeSelector,
-      ...annotationData,
-    };
-  } else {
-    podSpecOptions = {
-      resources: hardwareProfile.formData.resources,
-      tolerations: hardwareProfile.formData.selectedProfile?.spec.scheduling?.node?.tolerations,
-      nodeSelector: hardwareProfile.formData.selectedProfile?.spec.scheduling?.node?.nodeSelector,
-      ...annotationData,
-    };
-  }
-
-  return {
-    modelSize: {
-      sizes,
-      selectedSize: modelSize,
-      setSelectedSize: setModelSize,
-    },
-    hardwareProfile,
-    podSpecOptions,
-  };
-};
